Move EmailDAO
