+++ tachikoma-database/src/main/kotlin/com/sourceforgery/tachikoma/database/dao/EmailDAO.kt
@@ -1,5 +1,6 @@
-package com.sourceforgery.tachikoma.database.objects
+package com.sourceforgery.tachikoma.database.dao
 
+import com.sourceforgery.tachikoma.database.objects.EmailDBO
 import com.sourceforgery.tachikoma.identifiers.EmailId
 import io.ebean.EbeanServer
 import javax.inject.Inject
